MetaAnalysis.jl

Function Documentation

Main.MetaAnalysis.gIndGrps — Function

gIndGrps(m1,m2,s1,s2,n1,n2)

Hedge's g for independent groups.

Main.MetaAnalysis.vgIndGrps — Function

vgIndGrps(m1,m2,s1,s2,n1,n2) 

Variance estimates around the Hedge's g

Missing docstring.

Missing docstring for fixedEffects. Check Documenter's build log for details.

Missing docstring.

Missing docstring for tausq. Check Documenter's build log for details.

Missing docstring.

Missing docstring for randEffects. Check Documenter's build log for details.

Main.MetaAnalysis.forestPlot — Function

forestPlot(study, y, v) 

Forest Plot of the meta analysis.

Missing docstring.

Missing docstring for isq. Check Documenter's build log for details.

Missing docstring.

Missing docstring for metareg. Check Documenter's build log for details.

Main.MetaAnalysis.metaplot — Function

metaplot(y,v,x) 

Metaregression of point estimate on ONE variable 

```
# not run
y = df.y # where y is the point estimate of study in data frame df
v = df.v # where v is the variance around the point estimate
x = df.x # where x is a single variable
metaplot(y,v,x) # produces the scatterpot with linear predictor
```

Main.MetaAnalysis.funnel — Function

funnel(y,v)

Funnel Plot.

```
# not run
y = df.y
v = df.v
funnel(y,v) # produces the funnel plot
```

Main.MetaAnalysis.convert_or_d — Function

convert_or_d(logor, vlogor)

convert logoddsratio to d

Main.MetaAnalysis.convert_d_or — Function

convert_d_or(d,vd)

Convert from d to Odds Ratio

Main.MetaAnalysis.r_to_d — Function

r_to_d(r,vr)

Convert from correlation coefficient to Cohen's d

Missing docstring.

Missing docstring for r_to_z. Check Documenter's build log for details.

Main.MetaAnalysis.d_to_r — Function

d_to_r(d,vd,n1,n2)

Convert from Cohen's d to correlation coefficient

Main.MetaAnalysis.find_var_z — Function

find_var_z(n)

Find the variance of z statistic given sample size n

Missing docstring.

Missing docstring for turn_z2r. Check Documenter's build log for details.